requirements.txt
# Dependências do Python
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai>=0.7.1
# ollama==0.2.0  # Opcional: descomente para usar modelos locais do Ollama
